refactor(#3): move user BC to member BC folder
