+++ src/main/java/com/example/leaflog/bc/member/user/infrastructure/security/oauth/dto/OauthUserInfo.java
@@ -1,0 +1,8 @@
+package com.example.leaflog.bc.member.user.infrastructure.security.oauth.dto;
+
+public record OauthUserInfo(
+        String name,
+        String email,
+        String profileImg
+) {
+}
