+++ src/main/java/com/example/leaflog/bc/member/user/infrastructure/security/oauth/dto/TokenResponseDto.java
@@ -1,0 +1,10 @@
+package com.example.leaflog.bc.member.user.infrastructure.security.oauth.dto;
+
+public record TokenResponseDto(
+        String accessToken,
+        String refreshToken
+) {
+    public static TokenResponseDto of(String accessToken, String refreshToken){
+        return new TokenResponseDto(accessToken, refreshToken);
+    }
+}
