+++ src/main/java/com/example/leaflog/bc/member/user/infrastructure/security/oauth/handler/OauthFailureHandler.java
@@ -1,0 +1,24 @@
+package com.example.leaflog.bc.member.user.infrastructure.security.oauth.handler;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
+import org.springframework.stereotype.Component;
+import org.springframework.web.servlet.HandlerExceptionResolver;
+
+@Component
+@RequiredArgsConstructor
+public class OauthFailureHandler extends SimpleUrlAuthenticationFailureHandler {
+    private final HandlerExceptionResolver handlerExceptionResolver;
+
+    @Override
+    public void onAuthenticationFailure(
+            HttpServletRequest request,
+            HttpServletResponse response,
+            AuthenticationException exception
+    ) {
+        handlerExceptionResolver.resolveException(request, response, null, exception);
+    }
+}
