+++ src/main/java/com/example/leaflog/bc/member/user/infrastructure/security/oauth/parse/GithubOauthUserInfoParser.java
@@ -1,0 +1,28 @@
+package com.example.leaflog.bc.member.user.infrastructure.security.oauth.parse;
+
+
+import com.example.leaflog.bc.member.user.infrastructure.security.oauth.dto.OauthUserInfo;
+import lombok.RequiredArgsConstructor;
+
+import java.util.Map;
+
+@RequiredArgsConstructor
+public class GithubOauthUserInfoParser implements OauthUserInfoParser {
+
+    private final GithubEmailFetcher emailFetcher;
+
+    @Override
+    public OauthUserInfo parse(Map<String, Object> attributes, String accessToken) {
+
+        String name = attributes.get("name") != null
+                ? (String) attributes.get("name")
+                : (String) attributes.get("login");
+
+        String email = emailFetcher.fetchPrimaryEmail(accessToken);
+
+        String avatarUrl = (String) attributes.get("avatar_url");
+
+        return new OauthUserInfo(name, email, avatarUrl);
+    }
+}
+
