--- src/main/java/com/example/leaflog/bc/user/infrastructure/security/jwt/JwtProperties.java
@@ -1,13 +1,0 @@
-package com.example.leaflog.bc.user.infrastructure.security.jwt;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "spring.jwt")
-public record JwtProperties(
-        String secretKey,
-        Long accessExp,
-        Long refreshExp,
-        String header,
-        String prefix
-) {
-}
