--- src/main/java/com/example/leaflog/bc/user/infrastructure/security/jwt/JwtTokenProvider.java
@@ -1,63 +1,0 @@
-package com.example.leaflog.bc.user.infrastructure.security.jwt;
-
-import com.example.leaflog.bc.user.infrastructure.refreshtoken.RefreshToken;
-import com.example.leaflog.bc.user.infrastructure.refreshtoken.repository.RefreshTokenRepository;
-import com.example.leaflog.bc.user.infrastructure.security.auth.AuthDetails;
-import com.example.leaflog.bc.user.domain.User;
-import com.example.leaflog.bc.user.domain.repository.UserRepository;
-import com.example.leaflog.bc.user.domain.vo.GithubEmail;
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class JwtTokenProvider {
-    private final JwtProperties jwtProperties;
-    private final JwtTokenStructure jwtTokenStructure;
-    private final UserRepository userRepository;
-    private final RefreshTokenRepository refreshTokenRepository;
-
-    private static final String ACCESS_TOKEN = "access_token";
-    private static final String REFRESH_TOKEN = "refresh_token";
-
-    public String generateAccessToken(String email){
-        return jwtTokenStructure.generateToken(
-                email,
-                ACCESS_TOKEN,
-                jwtProperties.accessExp());
-    }
-
-    public String generateRefreshToken(String email){
-        String refreshToken = jwtTokenStructure.generateToken(
-                email,
-                REFRESH_TOKEN,
-                jwtProperties.refreshExp());
-
-        refreshTokenRepository.save(RefreshToken.builder()
-                .email(email)
-                .refreshToken(refreshToken)
-                .ttl(jwtProperties.refreshExp())
-                .build());
-        return refreshToken;
-    }
-
-    public String resolveToken(HttpServletRequest request){
-        String bearerToken = request.getHeader(jwtProperties.header());
-        return jwtTokenStructure.parseToken(bearerToken);
-    }
-
-    public Authentication getAuthentication(String token){
-        String email = jwtTokenStructure.getTokenSubject(token);
-        User user = userRepository.findByGithubEmail(GithubEmail.of(email))
-                .orElseThrow(() -> new UsernameNotFoundException("인증 실패"));
-
-        UserDetails userDetails = new AuthDetails(user);
-        return new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-    }
-}
-
