--- src/main/java/com/example/leaflog/bc/user/infrastructure/security/jwt/JwtTokenStructure.java
@@ -1,58 +1,0 @@
-package com.example.leaflog.bc.user.infrastructure.security.jwt;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.security.Keys;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.SecretKey;
-import java.nio.charset.StandardCharsets;
-import java.util.Date;
-
-@Component
-@RequiredArgsConstructor
-public class JwtTokenStructure {
-    private final JwtProperties jwtProperties;
-
-    private SecretKey getSecretKey(){
-        return Keys.hmacShaKeyFor(jwtProperties.secretKey().getBytes(StandardCharsets.UTF_8));
-    }
-
-    public String generateToken(String email, String type, Long exp){
-        return Jwts.builder()
-                .signWith(getSecretKey(), SignatureAlgorithm.HS256)
-                .setSubject(email)
-                .claim("type", type)
-                .setIssuedAt(new Date())
-                .setExpiration(new Date(System.currentTimeMillis() + exp * 1000))
-                .compact();
-    }
-
-    public String parseToken(String bearerToken){
-        if(bearerToken != null && bearerToken.startsWith(jwtProperties.prefix())){
-            return bearerToken.replace(jwtProperties.prefix(), "").trim();
-        }
-        return null;
-    }
-
-    public Claims getTokenBody(String token){
-        try{
-            return Jwts.parserBuilder()
-                    .setSigningKey(getSecretKey())
-                    .build()
-                    .parseClaimsJws(token)
-                    .getBody();
-        } catch (io.jsonwebtoken.ExpiredJwtException e){
-            throw new IllegalArgumentException("토큰 만료"); //ExpiredJwtException();
-        } catch (Exception e) {
-            throw new IllegalArgumentException("유효하지 않는  토큰"); //InvalidJwtException();
-        }
-    }
-
-    public String getTokenSubject(String token){
-        return getTokenBody(token).getSubject();
-    }
-}
-
