--- src/main/java/com/example/leaflog/bc/user/infrastructure/security/oauth/dto/OauthUserInfo.java
@@ -1,8 +1,0 @@
-package com.example.leaflog.bc.user.infrastructure.security.oauth.dto;
-
-public record OauthUserInfo(
-        String name,
-        String email,
-        String profileImg
-) {
-}
